Create a single AppTheme provider
This app theme provider handles the logic for the user switching PeterPortal themes as well as creating the appropriate MUI theme so that buttons (and presumably other components in the future) have the correct text color
