+++ site/src/App.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'react-bootstrap-range-slider/dist/react-bootstrap-range-slider.css';
@@ -16,77 +16,16 @@ import AdminPage from './pages/AdminPage';
 import ReviewsPage from './pages/ReviewsPage';
 import SideBar from './component/SideBar/SideBar';
 
-import ThemeContext from './style/theme-context';
-
 import trpc from './trpc';
-import { Theme } from '@peterportal/types';
 import { useAppDispatch } from './store/hooks';
 import { searchAPIResults } from './helpers/util';
 import { setCoursebag } from './store/slices/coursebagSlice';
 import { useIsLoggedIn } from './hooks/isLoggedIn';
-
-function isSystemDark() {
-  return window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches;
-}
+import AppThemeProvider from './component/AppThemeProvider/AppThemeProvider';
 
 export default function App() {
-  // default darkMode to local or system preferences
-  const [usingSystemTheme, setUsingSystemTheme] = useState(
-    localStorage.getItem('theme') === 'system' || !localStorage.getItem('theme'),
-  );
-  const [darkMode, setDarkMode] = useState(
-    usingSystemTheme ? isSystemDark() : localStorage.getItem('theme') === 'dark',
-  );
   const isLoggedIn = useIsLoggedIn();
-  const [prevDarkMode, setPrevDarkMode] = useState(false); // light theme is default on page load
   const dispatch = useAppDispatch();
-
-  /**
-   * we run this check at render-time and compare with previous state because a useEffect
-   * would cause a flicker for dark mode users on page load since the first render would be without
-   * the data-theme property set (light would be used by default)
-   */
-  if (darkMode != prevDarkMode) {
-    // Theme styling is controlled by data-theme attribute on body being set to light or dark
-    document.body.setAttribute('data-theme', darkMode ? 'dark' : 'light');
-    setPrevDarkMode(darkMode);
-  }
-
-  /**
-   * Sets the theme state
-   * @param theme
-   */
-  const setThemeState = useCallback((theme: Theme) => {
-    if (theme === 'system') {
-      setDarkMode(isSystemDark());
-      setUsingSystemTheme(true);
-    } else {
-      setDarkMode(theme === 'dark');
-      setUsingSystemTheme(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    const setSystemTheme = () => setThemeState('system');
-    const matcher = window.matchMedia('(prefers-color-scheme: dark)');
-
-    if (usingSystemTheme) matcher.addEventListener('change', setSystemTheme);
-    return () => matcher.removeEventListener('change', setSystemTheme);
-  }, [setThemeState, usingSystemTheme]);
-
-  /**
-   * Sets the theme state and saves the users theme preference.
-   * Saves to account if logged in, local storage if not
-   * @param theme
-   */
-  const setTheme = (theme: Theme) => {
-    setThemeState(theme);
-    if (isLoggedIn) {
-      trpc.users.setTheme.mutate({ theme });
-    } else {
-      localStorage.setItem('theme', theme);
-    }
-  };
 
   const loadCoursebag = useCallback(async () => {
     const courseIds = isLoggedIn
@@ -97,21 +36,12 @@ export default function App() {
   }, [dispatch, isLoggedIn]);
 
   useEffect(() => {
-    // if logged in, load user theme from db
-    if (isLoggedIn) {
-      trpc.users.get.query().then((res) => {
-        if (res.theme) {
-          setThemeState(res.theme);
-        }
-      });
-    }
-
     loadCoursebag();
-  }, [isLoggedIn, setThemeState, dispatch, loadCoursebag]);
+  }, [loadCoursebag]);
 
   return (
     <Router>
-      <ThemeContext.Provider value={{ darkMode, usingSystemTheme, setTheme }}>
+      <AppThemeProvider>
         <AppHeader />
         <div className="app-body">
           <div className="app-sidebar">
@@ -131,7 +61,7 @@ export default function App() {
           </div>
           <div className="changelog-modal">{<ChangelogModal />}</div>
         </div>
-      </ThemeContext.Provider>
+      </AppThemeProvider>
     </Router>
   );
 }
